Add an option to include datasets to user_show.
Now by default they are not added, so that in most cases there is no
need to be so heavy handed with the database. They can be added by
adding include_datasets to the data_dict as with groups.

Goes some way to fix #2130

Bonus bug fix: bulk_process in group controller was putting
include_datasets in the context instead of the data_dict.  This is now
fixed.